+++ src/utils.py
@@ -6,6 +6,8 @@ log_level = os.getenv('LOG_LEVEL', logging.ERROR)
 logger = logging.getLogger(__name__)
 logger.setLevel(log_level)
 
+DEFAULT_DB_INSTANCE_CLASS = 'db.r5.8xlarge'
+
 
 def describe_db_clusters(action):
     # Get all the instances
@@ -51,7 +53,7 @@ def purge_queue(queue_names):
         sqs.purge_queue(QueueUrl=queue_info['QueueUrl'])
 
 
-def disable_triggers(function_names):
+def disable_lambda_trigger(function_names):
     my_lambda = boto3.client('lambda', os.getenv('AWS_DEPLOYMENT_REGION', 'us-west-2'))
     return_value = False
     for function_name in function_names:
@@ -66,7 +68,7 @@ def disable_triggers(function_names):
     return return_value
 
 
-def enable_triggers(function_names):
+def enable_lambda_trigger(function_names):
     my_lambda = boto3.client('lambda', os.getenv('AWS_DEPLOYMENT_REGION', 'us-west-2'))
     return_value = False
     for function_name in function_names: